trunk_staging: add VanillaIceCream back to set RELEASE_PLATFORM_VERSION_ALL_CODENAMES
Include VanillaIceCream in the set of RELEASE_PLATFORM_VERSION_ALL_CODENAMES.

Commit generated by running the following commands:

    $ build-flag --quiet --release=trunk_staging set RELEASE_PLATFORM_VERSION_ALL_CODENAMES VanillaIceCream,Baklava

According to the description of RELEASE_PLATFORM_VERSION_ALL_CODENAMES

    A comma separated list of all codenames that are in preview. This
    variable is sort of a lie for historical reasons and only includes
    codenames up to and including the currently active codename

adding VanillaIceCream may or may not be the right thing to do. Submit
this for now to fix postsubmit errors (tracked in b/374457207).

Bug: 354023031
Bug: 374457207
Test: build-flag --quiet --all-releases get RELEASE_PLATFORM_VERSION_ALL_CODENAMES
Ignore-AOSP-First: unblock work to set codename to Baklava (this CL can be CP'd to AOSP afterwards)
Change-Id: I2fbc5d938cabc8dcdc64f5f21ab0c138f44a999f