Introduce RELEASE_BUILD_PURGE_PRODUCT_ADB_KEYS flag
If set it purges product adb key from image early during build. The
actual change to the .mk file is done in a separate commit.
The flag is set to true by default, and overwritten to false when build
variant is either userdebug, or eng across all release versions.
This way we keep the existing behavior on all trunks, and therefore this
commit is not indented to bring any functional chagnes.

BUG: b/373580227
Test: lunch husky-trunk_staging-{user,userdebug,eng} and check the
      contents of build_flags_vendor.json
Change-Id: Ie803ea0116b0de7b81b4d28aca841b95ae53a5b8